Refactor the code, change price to be FloatProperty, and change parent usage to be using parent_id

+++ models/package.py
@@ -4,6 +4,7 @@ from models.image_model import ImageModel
 
 
 class Package(ndb.Model):
+    destination_id = ndb.IntegerProperty()
     title = ndb.StringProperty()
     content = ndb.TextProperty()
     location = ndb.StringProperty()
@@ -13,9 +14,8 @@ class Package(ndb.Model):
     updated_date = ndb.DateTimeProperty(auto_now=True)
 
     @classmethod
-    def add(cls, destination, title, content, location, image, price):
-        parent = destination.key
-        package = cls(parent=parent, title=title, content=content, location=location)
+    def add(cls, destination_id, title, content, location, image, price):
+        package = cls(destination_id=int(destination_id), title=title, content=content, location=location)
         if image:
             image_model = ImageModel.add(image)
             package.image_id = image_model.key.id()
@@ -25,32 +25,35 @@ class Package(ndb.Model):
         return package
 
     @classmethod
-    def get_all_by_ancestor(cls, destination):
-        ancestor = destination.key
-        packages = cls.query(ancestor=ancestor).fetch()
+    def get_all_by_destination_id(cls, destination_id):
+        packages = cls.query(cls.destination_id == int(destination_id)).fetch()
         return packages
 
     @classmethod
-    def update(cls, destination, package_id, title, content, location, price):
-        parent = destination.key
-        package = cls.get_by_id(int(package_id), parent=parent)
+    def update(cls, destination_id, package_id, title, content, location, price):
+        package = cls.get_by_id(int(package_id))
         if package:
-            package.title = title
-            package.content = content
-            package.location = location
-            if price.isdigit():
-                package.price = float(price)
-            package.put()
-            return package
+            if package.destination_id == int(destination_id):
+                package.title = title
+                package.content = content
+                package.location = location
+                if price.isdigit():
+                    package.price = float(price)
+                package.put()
+                return package
+            else:
+                return False
         else:
             return False
 
     @classmethod
-    def delete(cls, destination, package_id):
-        parent = destination.key
-        package = cls.get_by_id(int(package_id), parent=parent)
+    def delete(cls, destination_id, package_id):
+        package = cls.get_by_id(int(package_id))
         if package:
-            package.key.delete()
-            return True
+            if package.destination_id == int(destination_id):
+                package.key.delete()
+                return True
+            else:
+                return False
         else:
             return False
